Unify config resolution logic
