+++ packages/remix-dev/config/findConfig.ts
@@ -1,0 +1,16 @@
+import path from "node:path";
+import fse from "fs-extra";
+
+export function findConfig(
+  dir: string,
+  basename: string,
+  extensions: string[]
+): string | undefined {
+  for (let ext of extensions) {
+    let name = basename + ext;
+    let file = path.join(dir, name);
+    if (fse.existsSync(file)) return file;
+  }
+
+  return undefined;
+}
