+++ packages/remix-dev/vite/index.ts
@@ -2,12 +2,6 @@
 // don't need to have Vite installed as a peer dependency. Only types should
 // be imported at the top level, or code that doesn't import Vite.
 import type { ReactRouterVitePlugin } from "./plugin";
-export type {
-  BuildManifest,
-  Preset,
-  VitePluginConfig,
-  ServerBundlesFunction,
-} from "./plugin";
 
 export const vitePlugin: ReactRouterVitePlugin = (...args) => {
   let { reactRouterVitePlugin } =
